+++ lib/middleware/websocket/interactors/payloads/client_creation_payload.rb
