+++ lib/middleware/websocket/interactors/payloads/race_payload.rb
