+++ spec/lib/middleware/websocket/interactors/payloads/client_creation_payload_spec.rb
