+++ spec/lib/middleware/websocket/interactors/payloads/race_payload_spec.rb
